RW-5503 move external links above searchbar
