+++ ui/src/app/cohort-search/list-search-v2/list-search-v2.component.tsx
@@ -3,19 +3,17 @@ import * as React from 'react';
 import {Key} from 'ts-key-enum';
 
 import {domainToTitle} from 'app/cohort-search/utils';
-import {Button, Clickable, StyledAnchorTag} from 'app/components/buttons';
+import {Button, Clickable} from 'app/components/buttons';
 import {ClrIcon} from 'app/components/icons';
 import {TextInput} from 'app/components/inputs';
 import {TooltipTrigger} from 'app/components/popups';
 import {Spinner, SpinnerOverlay} from 'app/components/spinners';
-import {AoU} from 'app/components/text-wrappers';
 import {cohortBuilderApi} from 'app/services/swagger-fetch-clients';
 import colors, {colorWithWhiteness} from 'app/styles/colors';
 import {reactStyles, withCdrVersions, withCurrentWorkspace} from 'app/utils';
 import {triggerEvent} from 'app/utils/analytics';
 import {attributesSelectionStore, setSidebarActiveIconStore} from 'app/utils/navigation';
 import {WorkspaceData} from 'app/utils/workspace-data';
-import {environment} from 'environments/environment';
 import {CdrVersion, CdrVersionListResponse, CriteriaType, DomainType} from 'generated/fetch';
 
 const borderStyle = `1px solid ${colorWithWhiteness(colors.dark, 0.7)}`;
@@ -146,7 +144,6 @@ const styles = reactStyles({
     height: '100%',
     lineHeight: '0.75rem',
     verticalAlign: 'middle',
-    width: '50%'
   }
 });
 
@@ -230,11 +227,6 @@ export const ListSearchV2 = fp.flow(withCdrVersions(), withCurrentWorkspace())(
     showStandardResults() {
       this.trackEvent('Standard Vocab Hyperlink');
       this.setState({standardOnly: true});
-    }
-
-    get showDataBrowserLink() {
-      return [DomainType.CONDITION, DomainType.PROCEDURE, DomainType.MEASUREMENT, DomainType.DRUG]
-        .includes(this.props.searchContext.domain);
     }
 
     get checkSource() {
@@ -383,26 +375,6 @@ export const ListSearchV2 = fp.flow(withCdrVersions(), withCurrentWorkspace())(
             </div>}
             {domain === DomainType.DRUG && <div style={styles.drugsText}>
               Your search may bring back brand names, generics and ingredients. Only ingredients may be added to your search criteria.
-            </div>}
-          </div>
-          <div style={{...styles.helpText, textAlign: 'right'}}>
-            {domain === DomainType.DRUG && <div>
-              <StyledAnchorTag
-                 href='https://mor.nlm.nih.gov/RxNav/'
-                 target='_blank'
-                 rel='noopener noreferrer'>
-                Explore
-              </StyledAnchorTag>
-              &nbsp;drugs by brand names outside of <AoU/>.
-            </div>}
-            {this.showDataBrowserLink && <div>
-              Explore Source information on the&nbsp;
-              <StyledAnchorTag
-                 href={environment.publicUiUrl}
-                 target='_blank'
-                 rel='noopener noreferrer'>
-                Data Browser.
-              </StyledAnchorTag>
             </div>}
           </div>
         </div>
